fix: performance

+++ app/src/index.ts
@@ -1,62 +1,55 @@
-import { Elysia, t } from "elysia";
 import { redisService } from "./services/RedisService";
 import { startWorker } from "./worker";
 import { paymentService } from "./services/PaymentService";
 
-const app = new Elysia()
-
-app.post('/payments', (ctx) => {
-  const { correlationId, amount } = ctx.body;
-
-  if (!correlationId) {
-    ctx.set.status = 400;
-    return { error: 'correlationId is not valid.' }
-  }
-
-  if (!amount) {
-    ctx.set.status = 400;
-    return { error: 'amount is not valid.' }
-  }
-
-  redisService.addToQueue(correlationId, amount);
-
-  ctx.set.status = 202
-  return { status: 'ok' }
-}, {
-  body: t.Object({
-    correlationId: t.String(),
-    amount: t.Number()
-  })
-})
-
-app.get('/payments-summary', async (ctx) => {
-  const { from, to } = ctx.query;
-
-  try {
-    const summary = await paymentService.getPaymentsSummary(from, to);
-
-    ctx.set.status = 200
-    return summary;
-
-  } catch (error) {
-
-    ctx.set.status = 500;
-    return { error: 'Failed to fetch payment summary.' };
-  }
-});
-
-app.post('/purge-payments', async (ctx) => {
-
-  await paymentService.purgePayments();
-
-  ctx.set.status = 200
-  return { status: 'ok' };
-});
-
 const PORT = process.env.PORT || 3333;
-app.listen(PORT, () => {
-  console.log(
-    `🦊 Elysia is running at ${app.server?.hostname}:${app.server?.port}`
-  );
-  startWorker()
+
+const server = Bun.serve({
+  port: PORT,
+  maxRequestBodySize: 1024,
+  async fetch(req) {
+    if (req.method === "POST" && req.url.includes("/payments")) {
+      const body = await req.text()
+      const { amount, correlationId } = JSON.parse(body)
+
+      redisService.addToQueue(correlationId, amount)
+
+      return new Response(null, { status: 202 })
+    }
+
+    if (req.method === "GET" && req.url.includes("/payments-summary")) {
+      try {
+        const url = new URL(req.url)
+        const from = url.searchParams.get("from") || undefined
+        const to = url.searchParams.get("to") || undefined
+
+        const summary = await paymentService.getPaymentsSummary(from, to);
+
+        return new Response(JSON.stringify(summary), {
+          status: 200,
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+
+      } catch (error) {
+        return new Response("Failed to fetch payment summary.", { status: 500 });
+      }
+    }
+
+    if (req.method === "POST" && req.url.includes("/purge-payments")) {
+      try {
+        await paymentService.purgePayments();
+        return new Response(JSON.stringify({ status: 'ok' }), { status: 200 })
+      } catch (error) {
+        return new Response("Failed to purge payments.", { status: 500 });
+      }
+    }
+
+    return new Response("Not found", { status: 404 })
+  },
+
 });
+
+console.log(`Happy happy happy: ${server.port}`)
+startWorker()
